+++ CHANGELOG.md
@@ -1,7 +1,16 @@
 # Change Log
 All notable changes to the "pivotaly" extension will be documented in this file.
 
-Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how to structure this file.
+#### [Unreleased]
+- Support for workspaces that are not git repos
+- Support for workspaces with multiple git repos
 
-## [Unreleased]
+## [0.0.2]
+#### Added
+- Error message when not in a git repo
+
+#### Fixed
+- Doesn't register different project IDs when switching workspaces 
+
+## [0.0.1]
 - Initial release